Task4 - AZURE LANGUAGE SERVICES/PII Recognition Application - Task4/requirements.txt
azure-ai-textanalytics
azure-core==1.23.0
fastapi
uvicorn

